requirements.txt
asyncio==3.4.3 
fastapi==0.109.0
langchain==0.1.0
langchain-openai==0.0.2
langchainhub==0.1.14
numpy==1.26.2
openai==1.7.2
opentelemetry-api==1.22.0
pydantic==2.5.1
uvicorn==0.25.0
python-dotenv
SpeechRecognition